Create last_digit_even_odd_checker.c
This C program takes an integer input from the user and checks the last digit of the number to determine if it's even or odd. It specifically identifies if the number ends in 0, 2, 4, 6, or 8 (even digits), otherwise it reports the number as odd and prints the last digit. It's a basic example of using the modulo operator and conditional statements.