Update prime_numbers_up_to_n.c
"This code in the file really challenged me a lot. I was just starting to learn while loops, and this code felt like an evolution for me. Just like Ülker takes a leap forward, this code was a major breakthrough for me in programming. It was truly tough, thank you. This code will hold a special place for me, so I’m setting it apart from my other codes."