Create fibonacci_sequence.c
A C program that generates and prints the Fibonacci sequence up to a user-specified number of elements using a for loop. 
In the Fibonacci sequence, each term is the sum of the two preceding ones, starting from 1 and 1. 
The program also displays the value of the last term in the sequence.