Create full_pyramid_pattern.c
A C program that prints a full pyramid pattern of asterisks (*) using nested for loops. 
The pyramid has a user-defined number of rows, with spaces added to center-align the stars. 
Currently, the program prints odd numbers of stars per row (1, 3, 5...), but it can be modified to print even numbers as well.